Revise `ScriptProcessor` logic (#14)
Each instance of `ScriptProcessor` is meant for one output script. This commit revises the logic to acommodate this. However, since the new RNKC LOKAD protocol requires more than one output script to be valid, an `addScript` method is added to allow higher level applications to supplement the processor with additional required data where required.

Also, a constant is added for op codes. This object will be filled out
as new op codes are supported in the library.