Add proper support for bootstrap nodes (#58)
This commit ensures that the bootstrap nodes for `P2PCoordinator` are
properly set up to handle bootstrapping.

Along with this change comes changes that revert previous fixes for the
`passthroughMapper`. To ensure the 3-phase example still works, we
update the example to directly use localhost IP address to listen.

We also add a new script that can be used to generate keypairs for
persistent P2P `PeerID` configurations.